Clarify that **use** is a keyword
Readers unfamiliar with Rust may land on this chapter and not know that the **use** feature listed here is a keyword. Since it is written in bold like the other features but is different from any of the others, it may be worth detailing here that **use** is a keyword.